Code Clean.
