+++ Assets/Scripts/Combat/PlayerHitCollidersController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6dc46d0b32de14ee7b2725cd52378100
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
