+++ Assets/Scripts/Combat/PlayerProjectile.cs
@@ -4,9 +4,6 @@ using UnityEngine;
 
 namespace Game.Combat
 {
-    /// <summary>
-    /// Script to be put on projectile prefabs and then placed into the SO_EnemyClassStats ScriptableObject
-    /// </summary>
     public class PlayerProjectile : MonoBehaviour
     {
         [SerializeField] float projectileSpeed = 1f;
@@ -17,20 +14,11 @@ namespace Game.Combat
         private void Start()
         {
             transform.right = launchDirection;
-            //transform.right = target.transform.position - transform.position;
         }
         private void Update()
         {
-            //if(launchDirection == null) return;
-
             float step = projectileSpeed * Time.deltaTime;
-
-            // if(isHoming && !target.IsDead())
-            // {
-            //     transform.right = target.transform.position - transform.position;
-            // }
             transform.Translate(Vector3.right * step);
-            
         }
         public void SetTarget(Vector3 launchDirection, float damage)
         {
@@ -42,8 +30,6 @@ namespace Game.Combat
         private void OnTriggerEnter2D(Collider2D other) 
         {
             if(other.GetComponent<EnemyHealth>() == null) return;
-            // if(target.IsDead()) return;
-
             other.GetComponent<EnemyHealth>().TakeDamage(damage);
             Destroy(gameObject);
         }
